minor trello fixes

+++ src/Nethermind/Nethermind.Facade/Simulate/SimulateBlockTracer.cs
@@ -60,7 +60,7 @@ public class SimulateBlockTracer : BlockTracer
 
         result.Calls.ForEach(callResult =>
         {
-            if (callResult.Type == ResultType.Success)
+            if (callResult.Type == (short)ResultType.Success)
             {
                 callResult.Logs?.ForEach(log =>
                 {
